Make including a namespace optional.
If the namespace key is left out of the qubernetes.yaml config
then don't create a namespace, and don't add a namespace to the
various kubernetes resources.
Creating the resource in a particular namespace can always be
done by passing in the namespace at creation time:
`kubectl apply --namespace=development -f out`.