Helper scripts to create and take down tunnels.
Set up tunnels for a minikube quorum deployment, so that applications, such as cakeshop,
can connected to the transaction manager without mapping proxied IP addresses,
e.g. minikube sets up proxy IPs, but these will be different from the IPs used by the internal
cluster, mainly in curl http://$IP:9001/partyinfo.

Does not prompt to verify the authenticity of the host and adds the host to known_hosts on the first connection.