Insert a valid key in `privateFor` field of test private contract.
Obtaint the tm.pub key of node2 and set it in the `privateFor` field
of the private contract so that it will work with out manually updating
it once the cluster is up.  Assumes that at least 2 nodes are deployed,
which is a reasonable assumption as 3 is the minium size of a raft deployment
and 7 is the minimum size of an Istanbul deployment.