Wait for quorum node to be healthy when adding new nodes.
When adding a new node, instead of waiting an arbitrary 10 seconds,
test `geth-exec` to see if quorum is healthy on the node that ran
the config updates, before proceeding to add the new nodes.
Especially important for raft: if the node which ran the config update,
is not healthy after 10 seconds, the correct raft additions will not be created.