Removing support for HostPath.
HostPath is not recommended and will not be deleted automatically when
the deployment is deleted, this causes issues for users. The recommended
persistence storage is Persistence Volume Claims. Removing all
references to HostPath storage, and the logic for setting HostPath as a
potential storage type. This also simplifies maintenance and testing.