Adding optional annotations on PVC.
Allow for configuring annotions on the K8s PVC generation, e.g.
volume.beta.kubernetes.io/storage-class, which can allow for setting
various storage classes, or other policies that the underlying k8s
cluster may require.

changing the config requirements in the qubernetes.yaml from
 storage:
    # PVC (Persistent_Volume_Claim - tested with GCP).
    Type: PVC
to
  storage:
    PVC:
      annotations:
        volume.beta.kubernetes.io/storage-class: "YOUR_STORAGE_CLASS"
      ## when redeploying cannot be less than previous values
      Capacity: 5Gi

as PVC is the only support storage for now. Staying backwards compatible
with the old configuration syntax, but this should be removed in the
future.