Add optional annotations & storageclass to PVCs. (#17)
* Added optional annotations on PVC.
* Added optional storageClassName for PVCs to config.
( (clean up) Removed dead code in quorum-deployment.yaml.erb which was left over from when hostPath was support. hostPath is no longer supported as a storage option, so remove logic around it.
---

* Added optional storageClassName for PVCs to config.
Allow the config to specific the storageClassName to use with the
PVC, currently both storageClassName and the annotation volume.beta.kubernetes.io/storage-class are supported, but the annotation volume.beta.kubernetes.io/storage-class won't be supported by Kubernetes in the future
see: https://kubernetes.io/docs/concepts/storage/persistent-volumes/#persistentvolumeclaims

An example of using storageClassName in the config file has been added here: examples/config/qubernetes-pvc-storageclass.yaml

storageClassName is optional, so if it is left out, the default for the cluster will be select by the underlying k8s platform.

* Added optional storageClassName for PVCs to config.
Allow for configuring annotations on the K8s PVC generation, e.g. volume.beta.kubernetes.io/storage-class, which can allow for setting various storage classes, or other policies that the underlying k8s cluster may require.

* changing the config requirements in the qubernetes.yaml from
```
 storage:
    # PVC (Persistent_Volume_Claim - tested with GCP).
    Type: PVC
```
to
```
  storage:
    PVC:
      annotations:
        volume.beta.kubernetes.io/storage-class: "YOUR_STORAGE_CLASS"
      ## when redeploying cannot be less than previous values
      Capacity: 5Gi
```
as PVC is now the only supported storage and `hostPath` support has been removed. For now, remain backwards compatible with the old configuration syntax, but this should be removed in the future.