Don't require geth entries to exist in config.
In the minimal configuration file, do not require geth attributes to
exit, if they do not exist, set defaults for geth parameters, e.g.
geth.network.id, geth.verbosity, geth rpc port, geth p2p2 port.

If geth.network.id isn't present in the config, default to chainId 1101, and if verbosity
isn't in the config, default to 9 - the highest debug level.

These values can be customized and set in the config if desired.