Add support websocket rcp, default to 8546.
Add support for connecting to geth via websockets. The default port is
set to 8546 and will always be set, but websockets need to be enabled in
the Geth_Startup_Param inside the config, e.g. --ws --wsorigins=\"*\" --wsaddr 0.0.0.0
see examples/config/qubernetes-ingress-ws.yaml

To override the default set it in the config unger `geth.Node_WSPort`.

note:
If trying to connect externally from an ingress be aware that depending
on the backend k8s service and how the load balancing is implenented the
health checks may fail, e.g. for GKE setting a configurable health check in k8s is not supported yet,
but will be available soon: https://github.com/kubernetes/ingress-gce/issues/1067.