Add support geth websockets rcp. (#28)
Add support for connecting to geth via websockets rpc. The default port is set to `8546` and will always be set, but in order to use websockets they need to be enabled in the `geth.Geth_Startup_Param` inside the config yaml, e.g. --ws --wsorigins=\"*\" --wsaddr 0.0.0.0 see examples/config/qubernetes-ingress-ws.yaml

To override the default set it in the config under `geth.Node_WSPort`.

note:
If trying to connect externally via an ingress be aware that depending on the backend k8s service and how the load balancing is implemented the health checks may fail, e.g. for GKE currently has a health check that uses http, and setting a configurable health check in k8s is not supported yet,
but will be available soon: https://github.com/kubernetes/ingress-gce/issues/1067.

* Change RPCPort 7nodes from 8546->8545.

change the 7nodes examples configs RPCPort to the default geth rpc port
so that it doesn't clash with the default websocket port (8546).