Move nodes into qubernetes.yaml for k8s generation.
* qubernetes qubernetes.yaml works with nodes inside qubernetes.yaml.
* fixed quorum generation to work with new node format, notes genesis
may have different consensus and chainId than the nodes, this is
supported for now, as that may be a true reflection of the network, but
this would be an incorrect network..
* remove permission-nodes.json.tmpl and script, as now it will be
updated with the hostnames when nodes are added. Use sym links to link
the file to the permissioned-nodes configMap dir, this is necessary,
because only configMaps mounted as directories will be updated without a
redeploy.