Switch peers from using variable IPs to hostnames.
* Use k8s service hostname for quorum peers: permissioned-nodes.json and static-nodes.json.

* Use k8s service hostname in tessera config for peer discovery.

* The hostnames are obtained via the quorum services which are known at the time of generation and do not change.
This removes the need to do sed and manipulation on the config files after they are deployed in order to set the IPs
(which are not known until after k8s resources have been deployed). This also helps with automatically updating the
config files when new nodes are deployed or removed, thus avoiding restarting the nodes.

* Enabled by this PR: https://github.com/jpmorganchase/quorum/pull/885.