Add quick-start-gen to generate a config.
Since the core configuration file no long support a simple number for
the number of nodes, but includes a node element add a quick-start-gen
to make getting started as easy as possible.

* Add a quick-start-gen which can generate a full config give a number
of nodes, consensus, and versions.
* Update quickest-start.sh so that it works with the quick-start-gen,
as updating the number of nodes in the core config is no longer
supported.