Change core config format to include explicit node blocks. (#59)
Getting rid of nodes.yaml and the simple specifying of the number of node in the core config. Having a simple way to specify the number of nodes, is good for easy of use, but is not a good for keeping the core config as a representation of the running state, and thus makes updating and day two ops tricky. The new config requires that the node blocks are explicitly included in the core config. There is a helper script for quick start `quick-start-gen` that can generate a base core config given the number of desired nodes and other optional params `quick-start-gen --help` the quickest-start.sh uses this, and works the same as before.


* Fix all quorum and k8s generators to work with new core config node format, *note* within the config, the genesis may have different consensus and chainId than the nodes, this is supported for now, as that may be a true reflection of the network, but this would be an incorrect network.

* Update all example config,s 7nodes, etc. to follow the new format.

 * Remove permission-nodes.json.tmpl and script, as now it will be updated with the hostnames when nodes are added. Use sym links to link  the file to the permissioned-nodes configMap dir, this is necessary, because only configMaps mounted as directories will be updated without a redeploy.

* Fixes and clean up of all templates.

* Updated README.md.