Initial cli qctl for interacting with qubernetes..
qctl is a command line client that can create qubernetes configs, deploy
the network, do some basic querying of the network and help manage nodes
(add/remove). The idea is to keep the qctl cli as modular as possible,
e.g. each command should do only one thing (Unix), allowing the commands
to be strung togehter in scripts, or leverage by UIs, etc.

This is a starting point, the main focus is to manipulate the base
quberetes config, which can be set as and ENV var (or passed in as a flag
to the commands) on the local / host machine, e.g. QUBE_CONFIG, the
K8s out directory, where the Kuberenetes resources and the quorum resource
have been generated to can allow be set as an EVN var, QUBE_K8S_DIR (or
passed in as a flag to the command that need them).

The basic initial flow to create and deploy a quorum Kubernetes network
(require docker and a running K8s platform - minikube, docker desktop,
kind), is:
> qctl init
> export QUBE_CONFIG=/Users/libby/Workspace.Quorum/qubernetes-priv/qubernetes.generate.yaml
> qctl generate network
> export QUBE_K8S_DIR=/Users/libby/Workspace.Quorum/qubernetes-priv/out
> qctl deploy network

Once the network is up and running you can query it (even before, the
queries run off the qubernetes config):
> qctl ls config
> qctl ls nodes --all

To add nodes:
> qctl add node --name=quorum-node5
> qctl generate network --update
> qctl deploy network

To connect / interact with nodes:
> qctl connect node1 quorum
> qctl geth attach node1
> qctl geth exec node1 "eth.blockNumber"

* Ported bash scripts to golang CLI qctl.
* See qctl --help for command list.
* Set up initial skaffolding for CLI and basic structure..
* Added namespace support for commands.
* Abstracted away dropping into a command.
* Allow for different version of the qubernetes container to be run,
default latest.
* Adding support for retrieving service info.
* Add type for URL (nodeport, clusterip).
* CLI allow multiple URL services to be returned.
* Set defaults for various values, e.g. --type  default clusterip, and
if no nodes are specified, look for all services containing "quorum".
* Include optional --nodeip flag for qctl get URL which will include the
k8s node ip in the output.
* Add create network and delete network, which will create and delete a
quorum K8s network on a running K8s platform: requiresn a k8s-dir flag be
passed in and that the directory exists.
* Set reasonable defaults for params / flags.
* Parse the qubernetes yaml.
* Load the yaml config when generating the network resources, and
display the network settings to the user.
* qctl init // initialized qubernetes config
* qctl generate network // generates network.
* Add README.md
* Remove prometheus log.
* Add `qctl ls config` which lists the currently used config.
* add ls node, add node.
* ls node optionally displays the fields of a node.
* Add node adds a new node to the qubernetes config, it does not add it
to the running network, that should be done with `qctl delploy network`
and perhaps additional steps (that can be scripted) will be required..
* When adding a node, if the user doesn't provide all the info,  get default
from the current qubernetes config.
* Show node enodeurl, create -> apply.
* When creating a network, use kubectl apply, not kubectl create.
* When displaying the node info, also optionally show enode id.